+++ src/components/Form/Form.format.tsx
@@ -156,7 +156,7 @@ const valueFormatters: FormDisplayValueFormatterMap<ReactNode> = {
   combobox: displayComboBoxValue,
   "icon-picker": displayIconPickerValue,
   search: displaySearchValue,
-  "type-search": String,
+  "search-text": String,
 };
 
 /**